Move Rules.withLogger scoped helper to the test target
`withLogger` is only used by tests, and the get/set/restore pattern is
short enough that it doesn't need to live in production code. Drop it
from `Rules` and add an equivalent extension under
`Tests/RulesEngineTests/Helpers/Rules+WithLogger.swift` so the one test
that wants a scoped override (`EvaluatorTests.testMissingVariable…`)
keeps its current call syntax.

`Evaluator.swift`'s doc comment is updated to drop the `withLogger`
mention.
